Shifted how things were being drawn
Shifted how things were being drawn, by moving most things that were
being drawn to their own canvas elements.

Doing this decreased the number of times the drawing api was used to
draw rectangles and choosing color styles, both of which are expensive.

Noticed a nice increase in reactivity to keyboard events as a perk as
well